Extract PDF import marker mapping
